+++ bosesoundtouchapi/models/languagecodes.py
@@ -35,3 +35,44 @@ class LanguageCodes(Enum):
     SLOVENIAN = 23
     TURKISH = 24
     HUNGARIAN = 25
+
+
+    @classmethod
+    def value_from_name(
+        cls,
+        name:str, 
+        default:int|None = None,
+        ) -> int | None:
+        """
+        Return a language value for its name (or value).
+
+        Args:
+            name (str):
+                Name to resolve; case-insensitive match will be performed.
+            default (int):
+                Default value to return if `name` argument could not be resolved.
+
+        Returns:
+            A value that represents the given name if found; 
+            otherwise, the default value.
+        """
+        if (not isinstance(name, str)):
+            return default 
+
+        # prepare for comparison.
+        name = name.upper()
+
+        # check for name match.
+        for member in LanguageCodes:
+            if member.name == name.upper():
+                return member.value
+
+        # check for value match.
+        for member in LanguageCodes:
+            if str(member.value) == name:
+                return member.value
+
+        # no match; return default value.
+        if (isinstance(default, cls)):
+            return default.value
+        return default